Move locate to its own module
