+++ ichnaea/service/geolocate/views.py
@@ -7,7 +7,7 @@ from ichnaea.service.error import (
     preprocess_request,
 )
 from ichnaea.service.base import check_api_key
-from ichnaea.service.locate import (
+from ichnaea.locate.locate import (
     PositionSearcher,
     map_data,
 )
